+++ src/_data/nav.js
@@ -1,0 +1,16 @@
+'use strict'
+
+const nav = {}
+
+nav.items = [
+  {
+    title: 'Mailer',
+    slug: './'
+  },
+  {
+    title: 'About',
+    slug: './about'
+  }
+]
+
+module.exports = nav