+++ src/app/app.11tydata.js
@@ -1,0 +1,9 @@
+'use strict'
+
+module.exports = function () {
+  return {
+    permalink: './{{ slug }}/index.html',
+    contentType: 'app',
+    layout: 'app'
+  }
+}