+++ src/emails/emails.11tydata.js
@@ -1,0 +1,9 @@
+'use strict'
+
+module.exports = function () {
+  return {
+    permalink: './email/{{ slug }}/index.html',
+    contentType: 'email',
+    layout: 'email'
+  }
+}