+++ CHANGELOG.md
@@ -16,4 +16,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## 0.1.0 - XXXX-XX-XX - XXXXXX
-- Add repo files
+- Add project repo files